triggers: Fix Trigger errors when executed
The Directive AFTER on trigger creation defines the new defined trigger
will be executed after any changes on a given table. In this way
the changes were firstly applied and after that the trigger was invoked.
The checks in the trigger were built on the basis that the changes
weren-t applied before the trigger invocation. To solve this issue,
it was neccessary to change AFTER with BEFORE keyword.

Signed-off-by: samanjot <singhamanjot726@gmail.com>